+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.2.1] - 2019-04-08
+### Added
+- Expose complete and by language builtin entities json configuration retrieval API [#13](https://github.com/snipsco/snips-nlu-parsers/pull/13)
+
+### Changed
+- Bump `snips-nlu-ontology` to `0.64.6` [#16](https://github.com/snipsco/snips-nlu-parsers/pull/16)
+
 ## [0.2.0] - 2019-02-27
 ### Changed
 - bump `snips-nlu-ontology` to `0.64.4`
@@ -16,6 +23,7 @@ All notable changes to this project will be documented in this file.
 - bump `snips-nlu-ontology` to `0.63.0`
 - re-export `gazetteer-entity-parser` crate
 
+[0.2.1]: https://github.com/snipsco/snips-nlu-parsers/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/snipsco/snips-nlu-parsers/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/snipsco/snips-nlu-parsers/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/snipsco/snips-nlu-parsers/compare/0.1.0...0.1.1